+++ src/common/components/form/Search/index.js
@@ -1,0 +1,2 @@
+export * from './SearchControlled';
+export * from './SearchField';